1.4.3.2-Leisure Upgrade
- Upgrade hashcounter to int64_t (IE allow big numbers in getmininginfo)
- Port sendmanyxml RPC command into Evo (allow pools to send-many txs),
and port serialize-legacy in (allow support sites to create proposal
trigger data)
- Added disableclientsidetransmission=1 key to disable any GSC
transmission
- Fixed QT txlist double click crash
- Added network m/hashps to getmininginfo

+++ src/validation.h
@@ -173,7 +173,7 @@ extern bool fCommunicatingWithPool;
 extern int64_t nLastDCContractSubmitted;
 extern int64_t nHPSTimerStart;
 extern int64_t nProposalStartTime;
-extern int nHashCounter;
+extern int64_t nHashCounter;
 extern double dHashesPerSec;
 extern bool fProposalNeedsSubmitted;
 extern int nProposalPrepareHeight;
